requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asyncpg==0.29.0
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
kombu==5.3.5
magic-filter==1.0.12
multidict==6.0.5
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
